+++ RELEASE_NOTES.md
@@ -1,3 +1,10 @@
+#### 0.10.0-alpha2 - November 24, 2017
+- SwaggerParser moved to runtime assembly (and exposed as API to library users) - [#90](https://github.com/fsprojects/SwaggerProvider/pull/90) 
+- Added runtime dependency on YamlDotNet
+- Removed dependency on FSharp.Data/JsonValue (replaced by JSON.NET)
+- Added full-init constructor for provided types [#91](https://github.com/fsprojects/SwaggerProvider/pull/91)
+- Option/Nullable types for not required properties [#93](https://github.com/fsprojects/SwaggerProvider/pull/93)
+
 #### 0.9.0 - November 17, 2017
 * Migration to new Type Provider SDK [#88](https://github.com/fsprojects/SwaggerProvider/pull/88)